### Bump tests
The behaviour of the system is simulated to observe the response to a bump in the road. The bump in the simulation below is 5 cm half sinusoid with a duration of half a second.

![[passive_suspension-bump_response.svg]]

The passive suspension dose slightly reduce the impulse but results in a larger displacement than the original bump and oscillates several times.

![[passive_suspension-bump_response-x_b.svg]]

![[passive_suspension-bump_response-all_props.svg]]

]]