### Initial Condition tests %%[[2024-12-18]] @ 15:08%%
The system is simulated with an initial velocity in the body and then in the wheel to compare how it recovers from a disturbance in each case.
The simulations where created using the [[initial_velocity_response.m]] script that can be found [here](https://github.com/jasht1/Uni-Projects/blob/master/State%20Space%20Control/CourseWork/code/initial_velocity_response.m) in the working directory.

![[passive_suspension-V_i_response.svg]] *Figure n* ^figure-n

The system recovers from disturbances to the wheel much faster than a similar disturbance to the body as can be seen in [[#^figure-n|figure n]]. However this is to be expected as at the same velocity the car body represents far more inertia and thus more energy for the system to dissipate.