### Initial Condition Test %%[[2024-12-19]] @ 01:37%%
The system is simulated with an initial velocity in the body and then in the wheel to compare how it recovers from a disturbance in each case.
The simulations where created using the [[initial_velocity_response.m]] script that can be found [here](https://github.com/jasht1/Uni-Projects/blob/master/State%20Space%20Control/CourseWork/code/initial_velocity_response.m) in the working directory.

![[active_suspension-V_i_response.svg]]

The initial wheel velocity of $0.5 \ m/s$ results in a very large acceleration of the car body. The acceleration is very short and results in negligible actual velocity or displacement of the car body but more than likely exceeds the maximum force the actuator can exert.