# Moving Reading Notes to Zotero %%[[2025-01-29]] @ 23:24%%
It has become clear that my current method for collecting resources and making notes on them is limiting me, it is far too manual and it is already becoming messy (see state at time of writing on the [git repo](https://github.com/jasht1/Uni-Projects/tree/5a763948d18894f3ff07c82c9aa55117cd79c660/Individual%20project/Notes/Reading%20Notes)). I have decided to dedicate some time to improving my workflow in the hopes that it will pay dividends if not in the course of this project then in due time.

Zotero looks a more sensible tool for managing sources and potentially reading notes, however this is yet to be seen.

[See Public Gist for details](https://gist.github.com/jasht1/1a0a0809755b3e28722fc2f0782b3efb#file-movingreadingnotestozotero-md) on my implementation.

![MovingReadingNotesToZotero.md](https://gist.github.com/jasht1/1a0a0809755b3e28722fc2f0782b3efb#file-movingreadingnotestozotero-md)

This is part of my ongoing mission to improve my workflow for ![[Academic writing]].